requirements.txt
dash==2.5.1
dash_bootstrap_components==1.1.0
numpy==1.22.3
pandas==1.4.2
plotly==5.8.2
pytest==6.2.5
PyYAML==6.0
scikit_learn==1.1.1
SQLAlchemy==1.4.32
flask==2.1.3